backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg-binary==3.2.12
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
sentence-transformers
torch>=2.2.0
pdfplumber==0.10.3
beautifulsoup4==4.12.3
google-cloud-storage==2.14.0
minio==7.2.3
tenacity==8.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.0.0
aiohttp
greenlet
